Fix: Put everything back into double_t
Because NEON loads from a pointer to int32[4] it assumes that each of the values are spaces sizeof(int32) apart. Thus casting a fixed_point_t to a fixed_point_double_t in the matrix doesn't work. Unless we can find a way to load into int32x4 from int16_t[4], we'll need to have all of our matricies be fixed_point_double_t